// src/type/type.ts
export interface propsType {
    isWidthLarge: boolean;
    isWidthSmall: boolean;
}
export type props = {
    scrollPosition?: boolean;
    isWidthLarge?: boolean;
    isWidthSmall?: boolean;
    isView?: boolean;
    isContentView?: boolean;
    bg?: string;
    slide?: boolean;
};
export type scroll = {
    Top: boolean;
    AboutMe: boolean;
    Skill: boolean;
    Project: boolean;
};
export type about = {
    first: boolean;
    second: boolean;
};
export type animationNumber = {
    first?: boolean;
    second?: boolean;
    third?: boolean;
    fourth?: boolean;
    fifth?: boolean;
    sixth?: boolean;
};